Server
So I figured this out a while ago, but forgot to upload it. Somehow a
few of the node_modules file folders got deleted, so here they are. This
is just for the server, but I just finished working on the css styling
and fixing the layouts so they look nicer, Tiara is working on the
LeaderBoard and we should hopefully have some more stuff upload in a
bit!